+++ llarp/tooling/rc_event.hpp
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <tooling/router_event.hpp>
 
 #include <router_contact.hpp>
@@ -7,13 +9,23 @@ namespace tooling
   struct RCGossipReceivedEvent : public RouterEvent
   {
     RCGossipReceivedEvent(const llarp::RouterID& routerID,
-                          const llarp::RouterContact& rc);
+                          const llarp::RouterContact& rc_)
+        : RouterEvent("RCGossipReceivedEvent", routerID, true), rc(rc_)
+    {
+    }
 
     std::string
-    ToString() const override;
+    ToString() const override
+    {
+      return RouterEvent::ToString()
+          + " ---- other RouterID: " + llarp::RouterID(rc.pubkey).ShortString();
+    }
 
     std::string
-    LongString() const;
+    LongString() const
+    {
+      return RouterEvent::ToString() + " ---- RC: " + rc.ToString();
+    }
 
     llarp::RouterContact rc;
   };
@@ -21,13 +33,23 @@ namespace tooling
   struct RCGossipSentEvent : public RouterEvent
   {
     RCGossipSentEvent(const llarp::RouterID& routerID,
-                      const llarp::RouterContact& rc);
+                      const llarp::RouterContact& rc_)
+        : RouterEvent("RCGossipSentEvent", routerID, true), rc(rc_)
+    {
+    }
 
     std::string
-    ToString() const override;
+    ToString() const override
+    {
+      return RouterEvent::ToString() + " ---- sending RC for RouterID: "
+          + llarp::RouterID(rc.pubkey).ShortString();
+    }
 
     std::string
-    LongString() const;
+    LongString() const
+    {
+      return RouterEvent::ToString() + " ---- RC: " + rc.ToString();
+    }
 
     llarp::RouterContact rc;
   };
