+++ llarp/tooling/router_event.hpp
@@ -28,12 +28,23 @@ namespace tooling
   struct RouterEvent
   {
     RouterEvent(std::string eventType, llarp::RouterID routerID,
-                bool triggered);
+                bool triggered)
+        : eventType(eventType), routerID(routerID), triggered(triggered)
+    {
+    }
 
     virtual ~RouterEvent() = default;
 
     virtual std::string
-    ToString() const;
+    ToString() const
+    {
+      std::string result;
+      result += eventType;
+      result += " [";
+      result += routerID.ShortString();
+      result += "] -- ";
+      return result;
+    }
 
     const std::string eventType;
